requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
fastapi==0.63.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pybase64==1.1.4
pycparser==2.20
pydantic==1.8.1
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dotenv==0.16.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
datetime==4.3
Authlib==0.15.3
